refactor(#615): carve worksource.rs into worksource/{mod,issues,prs}.rs -- pure move
Same line-coverage-asserted mechanism as the kanban carve: 29 contiguous
segments whose concatenation reproduces the original 1,541-line file
byte-for-byte, every line claimed exactly once. No logic changes.

- worksource/mod.rs (563): plugin discovery/cache, call_plugin + timeout
  machinery, the require_plugin/plugin_call fail-closed boundary,
  watch.toml config resolution + the infra tests (incl. the unix
  call_plugin_timeout suite).
- worksource/issues.rs (475): ExternalIssue/CreatedIssue, issue ops,
  comment, sync_issues, extract_issue_number + their regression guards.
- worksource/prs.rs (543): PR types and ops, check-state partition,
  job_id_from_link + their tests.

mod.rs re-exports issues::* and prs::* so every crate::worksource:: path
is unchanged; 40 declared scaffold lines (headers, mod/pub use, use decls,
test-module wrappers). cargo fmt deferred to its own commit.
